refacter ErrorPicker and BUFFER_PICKER
