+++ xds/src/test/java/io/grpc/xds/XdsLoadBalancerTest.java
@@ -21,7 +21,7 @@ import static io.grpc.ConnectivityState.CONNECTING;
 import static io.grpc.ConnectivityState.READY;
 import static io.grpc.ConnectivityState.TRANSIENT_FAILURE;
 import static io.grpc.LoadBalancer.ATTR_LOAD_BALANCING_CONFIG;
-import static io.grpc.xds.LocalityStore.LocalityStoreImpl.BUFFER_PICKER;
+import static io.grpc.xds.XdsSubchannelPickers.BUFFER_PICKER;
 import static junit.framework.TestCase.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
@@ -71,7 +71,7 @@ import io.grpc.internal.JsonParser;
 import io.grpc.internal.testing.StreamRecorder;
 import io.grpc.stub.StreamObserver;
 import io.grpc.testing.GrpcCleanupRule;
-import io.grpc.xds.LocalityStore.LocalityStoreImpl.ErrorPicker;
+import io.grpc.xds.XdsSubchannelPickers.ErrorPicker;
 import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
